Separate LinearSweep and LinearSweepParameter.
